Correct the Chunks implementation with the new type-aware one
